zephyr: sof/lib: split dma.h to dma.h and dma-legacy.h
Move parts of the dma.h interface only needed by XTOS drivers to a
separate dma-legacy.h file. Use of non-native drivers is a transitory
tool and used by a subset of targets. To make the code easier to read
and maintain, the main dma.h should only require definitions needed when
building SOF Zephyr with native Zephyr drivers
(CONFIG_ZEPHYR_NATIVE_DRIVERS). This is especially important for DMA as
there is lot of similarly named definitions and interfaces related to
DMA in Zephyr and SOF.
